requirements.txt
azure-ai-documentintelligence
azure-core>=1.30.0
azure-identity
azure-search-documents>=11.4.0
azure-storage-blob>=12.13.0
apscheduler>=3.10.0
beautifulsoup4
fastapi>=0.110.0
openai>=1.30.0
pymupdf
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pyodbc>=5.1.0
requests
streamlit
tiktoken
uvicorn[standard]>=0.29.0
